Move situation type variable to situation module
